Added the name of the restaurant in the “My reviews” section
NOTE: This is a breaking change. You’ll need to delete your data and re-populate it.

Realized the “My reviews” section wasn’t very useful, so I added the name of the original restaurant. Then put that in the table view instead of the user info, which isn’t really needed. This required denormalizing the restaurant name, which will have implications when you edit a restaurant.